% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_roulette.R
\name{agent_roulette_customs_visualise}
\alias{agent_roulette_customs_visualise}
\title{Create Agent Roulette Visualisation}
\usage{
agent_roulette_customs_visualise(
  attacking_nplayers = 5,
  defending_nplayers = 5,
  agents_to_exclude = NULL
)
}
\arguments{
\item{attacking_nplayers}{number of players on attacking team}

\item{defending_nplayers}{number of players on defending team}

\item{agents_to_exclude}{agents to exclude from the selection}
}
\value{
ggplot2
}
\description{
Create Agent Roulette Visualisation
}
\examples{
agent_roulette_two_teams_visualise()
}
