% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_roulette.R
\name{val_strat}
\alias{val_strat}
\title{Valorant Strats}
\usage{
val_strat(side = c("Both", "Attacking", "Defending"), seed = NULL, html = TRUE)
}
\arguments{
\item{side}{one of 'Both', 'Attacking', 'Defending'}

\item{seed}{random seed (numeric)}

\item{html}{should html formatting tags be added (flag)}
}
\value{
vector of strings containing valorant strats
}
\description{
Get a single random valorant strat
}
\examples{
val_strat()
}
