% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_roulette.R
\name{val_strats}
\alias{val_strats}
\title{Valorant Strats}
\usage{
val_strats(side = c("Both", "Attacking", "Defending"), seed = NULL)
}
\arguments{
\item{side}{one of 'Both', 'Attacking', 'Defending'}

\item{seed}{random seed (numeric)}
}
\value{
vector of strings containing valorant strats
}
\description{
Get vector of random valorant stats
}
\examples{
val_strats()
}
